Move callbacks to spearate file.
