+++ server.go
@@ -45,40 +45,6 @@ type TaipeionBot struct {
 	eventSemaphore     *semaphore.Weighted
 }
 
-// # Simple webhook event callback.
-//
-// This callback puts the incoming message to the log (stdout).
-func SimpleWebhookEventCallback(bot *TaipeionBot, event ChatbotWebhookEvent) error {
-	// Check if incoming event is text message.
-	if event.Message.Type != "text" {
-		log.Println("[SimpleStdCallback] Received non-text message. Ignoring.")
-		return nil
-	}
-
-	log.Printf("[SimpleStdCallback] Received event: %#v\n", event)
-	return nil
-}
-
-// # Private message callback.
-//
-// This callback is used to send a reply to a cer
-func PrivateMessageCallback(bot *TaipeionBot, event ChatbotWebhookEvent) error {
-	// Check if incoming event is text message.
-	if event.Message.Type != "text" {
-		log.Println("[PrivateMessageCallback] Received non-text message. Ignoring.")
-		return nil
-	}
-
-	reply_message := event.Message.Text
-	receiver := event.Source.UserId
-
-	// Process the incoming message.
-	reply_message = fmt.Sprintf("Bot: %s", reply_message)
-
-	return bot.SendPrivateMessage(receiver, reply_message)
-
-}
-
 // # Enqueue an incoming webhook event.
 func (tpb *TaipeionBot) enqueueWebhookIncomingEvent(event ChatbotWebhookEvent) {
 	tpb.eventQueue <- event
